重构 ShaderRegister 类, 修改 DynamicDescriptorHeap 以支持更多的描述符, 为 bindless 做准备, 新增加 ShadowPass 文件, 新增加 Shadow.hlsl shader, 添加 Pix 包

+++ Dx12Renderer/Dx12lib/Context/CommandList.cpp
@@ -539,7 +539,7 @@ void CommandList::clearDepthStencil(const DepthStencilView &dsv, float depth, UI
 }
 
 void CommandList::beginEvent(const std::string &eventName) {
-	_pCommandList->BeginEvent(1, eventName.c_str(), eventName.length()+1);
+	_pCommandList->BeginEvent(1, eventName.c_str(), static_cast<UINT>(eventName.length() + 1));
 }
 
 void CommandList::endEvent() {
